Refactor: Remove 5,850 lines of confirmed dead code
Delete 9 unreachable files:
- 7 dead screens (SimpleTeamScreen, EventsScreen, LeagueDetailScreen,
  MyTeamsScreen, CompetitionsListScreen, TeamsScreen, DonateScreen)
- 2 dead services (SimpleLeaderboardService, SimpleCompetitionService)

All verified dead by tracing navigation paths — zero navigate() calls
reach any of these routes. Redirect dangling EnhancedTeamScreen
navigation to ClubPage (the live club detail screen).

Extract CompetitionEvent type to src/types/competitionEvent.ts for
the 3 live event store services that imported it.

TypeScript compiles clean.
