+++ src/navigation/navigationHandlers.ts
@@ -106,91 +106,11 @@ export const createNavigationHandlers = (): NavigationHandlers => {
     handleTeamView: async (
       team: DiscoveryTeam,
       navigation: any,
-      userNpub?: string
+      _userNpub?: string
     ) => {
-      console.log(
-        'NavigationHandlers: Navigating to team dashboard:',
-        team.name
-      );
-
-      // Use passed userNpub (from working discovery page auth) instead of AsyncStorage lookups
-      let currentUserNpub: string | undefined = userNpub;
-
-      console.log(
-        '🔄 NavigationHandlers: User from passed parameter (same as working discovery):',
-        {
-          hasNpub: !!currentUserNpub,
-          npubSlice: currentUserNpub?.slice(0, 20) + '...' || 'undefined',
-        }
-      );
-
-      // Only try AsyncStorage fallback if no npub was passed
-      if (!currentUserNpub) {
-        try {
-          const userData = await AuthService.getCurrentUserWithWallet();
-          currentUserNpub = userData?.npub;
-
-          console.log('🔧 NavigationHandlers: Fallback to AuthService:', {
-            hasUser: !!userData,
-            hasNpub: !!currentUserNpub,
-            npubSlice: currentUserNpub?.slice(0, 20) + '...' || 'undefined',
-          });
-        } catch (error) {
-          console.error(
-            '❌ NavigationHandlers: Failed to get user from AuthService:',
-            error
-          );
-          // Final fallback to store
-          const user = useUserStore.getState().user;
-          currentUserNpub = user?.npub;
-          console.log('🔧 NavigationHandlers: Final fallback to store:', {
-            hasUser: !!user,
-            hasNpub: !!currentUserNpub,
-          });
-        }
-      }
-
-      // Use the same logic as EnhancedTeamScreen to determine membership
-      const calculatedUserIsMember = isTeamMember(currentUserNpub, team);
-
-      // Get captain status from cache (set by TeamCard where it works correctly)
-      let userIsCaptain = false;
-      if (team.id && currentUserNpub) {
-        const cachedStatus = await CaptainCache.getCaptainStatus(team.id);
-        if (cachedStatus !== null) {
-          userIsCaptain = cachedStatus;
-          console.log(
-            `✅ NavigationHandlers: Using cached captain status for ${team.name}: ${userIsCaptain}`
-          );
-        } else {
-          // Fallback only if not cached
-          userIsCaptain = isTeamCaptain(currentUserNpub, team);
-          console.log(
-            `⚠️ NavigationHandlers: No cached status, calculated: ${userIsCaptain}`
-          );
-          // Cache it for next time
-          await CaptainCache.setCaptainStatus(team.id, userIsCaptain);
-        }
-      }
-
-      // Member status includes both regular members and captains
-      const userIsMember = calculatedUserIsMember || userIsCaptain;
-
-      console.log('🎖️ NavigationHandlers: Team view navigation:', {
-        teamName: team.name,
-        userNpub: currentUserNpub?.slice(0, 8) + '...',
-        teamCaptainId:
-          'captainId' in team ? team.captainId?.slice(0, 8) + '...' : 'N/A',
-        userIsCaptain,
-        calculatedUserIsMember,
-        finalUserIsMember: userIsMember,
-      });
-
-      navigation.navigate('EnhancedTeamScreen', {
-        team,
-        userIsMember,
-        currentUserNpub, // Pass the working npub to avoid component-level AsyncStorage corruption
-        userIsCaptain, // Pass the correctly calculated captain status
+      navigation.navigate('ClubPage', {
+        clubId: team.id,
+        clubName: team.name,
       });
     },
